+++ src/accelerate/utils/__init__.py
@@ -72,5 +72,12 @@ if is_deepspeed_available():
 
 from .launch import PrepareForLaunch
 from .memory import find_executable_batch_size
-from .other import extract_model_from_parallel, get_pretty_name, patch_environment, save, wait_for_everyone
+from .other import (
+    extract_model_from_parallel,
+    get_pretty_name,
+    patch_environment,
+    save,
+    wait_for_everyone,
+    write_basic_config,
+)
 from .random import set_seed, synchronize_rng_state, synchronize_rng_states
